fix(pipeline,scan,llm): fix JSON type flip, dedup verdict loss, access-control FPs, docs drift
- AuditCost::toArray()['by_role'] flipped between a JSON array and object
  depending on which use case produced the report.
- AuditOrchestrator::isSameIdDuplicate() discarded a later iteration's
  validated verdict outright once any validated verdict already existed
  at the same id, even when it corrected the severity or type.
- RouteAttributeParser dropped a #[Route] path given in Symfony's
  locale-keyed array form, causing a false LACKS_ACCESS_CHECK.
- PhpParserControllerAccessControlParser missed a nullsafe
  denyAccessUnlessGranted() call, another false LACKS_ACCESS_CHECK.
- AnthropicTokenEstimator::supports() diverged from
  CostCalculator::isClaudeModel()'s substring check, contradicting a
  prior round's stated intent to match it.
- resources/schema.json and docs/configuration.md still documented
  audit.budget.max_cost_usd's old (> 0) minimum after a prior round
  tightened it to >= 0.01; the fast profile's own docs undersold its
  attacker concurrency.
- ProcessGitChangedFilesResolver::timeoutSeconds widened from int to
  float so its own timeout test can use a sub-second timeout instead of
  waiting a full second for a real process timeout.

+++ src/Audit/Infrastructure/Diff/ProcessGitChangedFilesResolver.php
@@ -59,7 +59,7 @@ final readonly class ProcessGitChangedFilesResolver implements GitChangedFilesRe
      */
     public function __construct(
         private Filesystem $filesystem = new Filesystem(),
-        private int $timeoutSeconds = self::DEFAULT_TIMEOUT_SECONDS,
+        private float $timeoutSeconds = self::DEFAULT_TIMEOUT_SECONDS,
         private ?Closure $gitDiffProcessFactory = null,
     ) {}
 
@@ -91,7 +91,7 @@ final readonly class ProcessGitChangedFilesResolver implements GitChangedFilesRe
         $process = new Process(['git', 'rev-parse', '--is-inside-work-tree'], $projectPath);
 
         try {
-            $process->setTimeout((float) $this->timeoutSeconds);
+            $process->setTimeout($this->timeoutSeconds);
             $process->run();
         } catch (ExceptionInterface $exception) {
             throw GitChangedFilesUnavailableException::forProcessFailure(\sprintf('determine whether "%s" is a git working tree', $projectPath), $exception);
@@ -108,7 +108,7 @@ final readonly class ProcessGitChangedFilesResolver implements GitChangedFilesRe
         $process = new Process(['git', 'rev-parse', '--verify', '--quiet', $ref], $projectPath);
 
         try {
-            $process->setTimeout((float) $this->timeoutSeconds);
+            $process->setTimeout($this->timeoutSeconds);
             $process->run();
         } catch (ExceptionInterface $exception) {
             throw GitChangedFilesUnavailableException::forProcessFailure(\sprintf('verify git ref "%s"', $ref), $exception);
@@ -149,7 +149,7 @@ final readonly class ProcessGitChangedFilesResolver implements GitChangedFilesRe
         $process = $factory($argv, $projectPath);
 
         try {
-            $process->setTimeout((float) $this->timeoutSeconds);
+            $process->setTimeout($this->timeoutSeconds);
             $process->mustRun();
         } catch (ProcessFailedException $processFailedException) {
             throw GitChangedFilesUnavailableException::fromProcessFailure($argv[\count($argv) - 1] ?? '', $process->getErrorOutput(), $processFailedException);
